COMP: Fix windows transforms module build error related to sqrt()
Following the update of ITK in r24360, the version of sqrt() function
accepting integer is probably not exposed anymore. This commit fixes
the following error by using an explicit cast:

// ---------------------
40>264>..\..\..\..\..\Slicer-1\Modules\Loadable\Transforms\qSlicerTransformsModuleWidget.cxx(447) : error C2668: 'sqrt' : ambiguous call to overloaded function

40>264>        C:\Program Files (x86)\Microsoft Visual Studio 9.0\VC\include\math.h(581): could be 'long double sqrt(long double)'
40>264>        C:\Program Files (x86)\Microsoft Visual Studio 9.0\VC\include\math.h(533): or       'float sqrt(float)'
40>264>        C:\Program Files (x86)\Microsoft Visual Studio 9.0\VC\include\math.h(128): or       'double sqrt(double)'
40>264>        while trying to match the argument list '(int)'
// ---------------------

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=24362
git-svn-id: http://svn.slicer.org/Slicer4/trunk@24362 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/Loadable/Transforms/qSlicerTransformsModuleWidget.cxx
@@ -444,8 +444,8 @@ void qSlicerTransformsModuleWidget::onPasteShortcutActivated()
                   "Unable to cast string to double: " << entry;
       return;
       }
-    tempMatrix->SetElement(index / static_cast<int>(sqrt(entries.count())),
-                           index % static_cast<int>(sqrt(entries.count())),
+    tempMatrix->SetElement(index / static_cast<int>(sqrt(static_cast<double>(entries.count()))),
+                           index % static_cast<int>(sqrt(static_cast<double>(entries.count()))),
                            numericEntry);
     ++index;
   }
